+++ RELEASE-NOTES.md
@@ -9,6 +9,7 @@
 * #312 DataFrame.get(col, row)
 * #313 Upgrade ECharts to 5.5.1
 * #316 ECharts AxisLabel options: rotate, font...
+* #317 ECharts: bar chart ignores "stack" setting
 
 ## Release 1.0.0-M22